moved constants , interfaces and types folder inside utils.
