+++ src/utils/constants/board-card.ts
