+++ src/utils/constants/board.ts
