+++ src/utils/validation.ts
@@ -1,6 +1,3 @@
-import bcrypt from "bcryptjs";
-import { getAllUsers } from "../services/auth";
-
 export function validatePassword(password: string): string | null {
   const minLength = 8;
   const uppercasePattern = /[A-Z]/;
@@ -39,18 +36,6 @@ export function validateEmail(email: string): string | null {
   return null;
 }
 
-
 export function normalizeEmail(email: string) {
   return email.trim().toLowerCase();
-}
-
-
-export function validateUser(email: string, password: string): boolean {
-  const users = getAllUsers();
-  const normalizedEmail = normalizeEmail(email);
-
-  const existingUser = users.find((user) => user.email === normalizedEmail);
-  if (!existingUser) return false;
-
-  return bcrypt.compareSync(password, existingUser.password);
 }
